# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API controle de acesso Laravel",
        "description": "Documentação com todos os endpoints necessários para a correta utilização da aplicação.",
        "contact": {
            "email": "assuncaofilipe97@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/app/forgot-password": {
            "post": {
                "tags": [
                    "Recuperação de senha"
                ],
                "summary": "Recuperar senha",
                "description": "Envia um código de recuperação ao email informado.",
                "operationId": "6e772c6a1b43cd8d8af73e05bcac449f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "register": {
                                    "summary": "Recuperar senha",
                                    "value": {
                                        "email": "exemplo@gmail.com"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código de recuperação enviado ao email.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "message": "Código de recuperação de senha enviado ao seu email."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Email inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "message": "E-mail não encontrado, favor revisar!"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/app/validate-password-token": {
            "post": {
                "tags": [
                    "Recuperação de senha"
                ],
                "summary": "Validar código de verificação",
                "description": "Válida código de verificação enviado ao email.",
                "operationId": "66aabca5a25c369e7588f20c5b336ffa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password_token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "password_token": {
                                    "summary": "Recuperar senha",
                                    "value": {
                                        "password_token": "z5xYKu"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código de verificação válido,\n                        retorna um novo código de verificação,\n                        que será utilizado como parâmetro na rota new-password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "password_token": "L025fa"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Código de verificação expirado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "message": "Código de verificação expirado."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Código de verificação inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "Código verificação inválido": {
                                        "summary": "Código verificação inválido",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "message": "Código de verificação inválido."
                                            }
                                        }
                                    },
                                    "Código de verificação maior que 6 caracteres": {
                                        "summary": "Maior do que 6 caracteres",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "message": "O campo password token não pode ser superior a 6 caracteres."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/app/new-password": {
            "post": {
                "tags": [
                    "Recuperação de senha"
                ],
                "summary": "Nova senha",
                "description": "Criação de nova senha.",
                "operationId": "4816c7f37cbfaf004c920ca91c530a07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password_token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "password_token": {
                                    "summary": "Nova senha",
                                    "value": {
                                        "password_token": "L025fa",
                                        "password": "123f123f",
                                        "password_confirmation": "123f123f"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "message": "Senha alterada com sucesso."
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Código de verificação expirado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "message": "Código de verificação expirado."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Código de verificação inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "Código de verificação inválido": {
                                        "summary": "Código de verificação inválido",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "message": "Código de verificação inválido."
                                            }
                                        }
                                    },
                                    "Código de verificação maior que 6 caracteres": {
                                        "summary": "Maior do que 6 caracteres",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "message": "O campo password token não pode ser superior a 6 caracteres."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}